daily-emailing/requirements.txt
python-dotenv
pandas
boto3
psycopg2-binary